Only check if ticket has expired before today
We will change so that when the admin is about to buy the showing
he or she will be shown if a particular gift certificate has
expired or not. Then the admin doesn't even have to try to use a
particular ticket since they know if it is expired already.