refactor: rework how we use the principal
Now we only fetch the ID token when we create a user on login. The
principal is created from the jwt token and included in the context.

The principal cache is removed since it isn't needed anymore.

We might need a way to update user info upon login since some important
info may have changed from the IDP. E.g. the mail could've been changed.