feat(login): make state stateless
Instead of generating the state and caching the verification variables
using said state in-memory, we encrypt it using AES-GCM and send it to
the authorization server.

This works since the state is passed back to us during the code
exchange. We can also reuse the nonce for encryption and CSRF protection
since the nonce is used to decrypt the state. If it were somehow
tampered with, we would be able to decrypt and verify the state.

This solves a denial-of-service attack where an attacker could spam
login requests without ever following the redirect URL thus exhausting
all available memory.